Update e2e to be part of existing front-end service
