--- e2e/extra-config-options/docker-mft/go.mod
@@ -1,5 +1,0 @@
-module github.com/aws/copilot-cli/e2e/extra-config-options/docker-mft
-
-go 1.14
-
-require github.com/julienschmidt/httprouter v1.3.0
