--- e2e/extra-config-options/extra_config_test.go
@@ -1,196 +1,0 @@
-// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package extra_config_test
-
-import (
-	"encoding/json"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"net/http"
-	"os"
-
-	"github.com/aws/copilot-cli/e2e/internal/client"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("extra config flow", func() {
-	Context("when creating a new app", func() {
-		var (
-			initErr error
-		)
-		BeforeAll(func() {
-			_, initErr = cli.AppInit(&client.AppInitRequest{
-				AppName: appName,
-			})
-		})
-
-		It("app init succeeds", func() {
-			Expect(initErr).NotTo(HaveOccurred())
-		})
-
-		It("app init creates an copilot directory and workspace file", func() {
-			Expect("./copilot").Should(BeADirectory())
-			Expect("./copilot/.workspace").Should(BeAnExistingFile())
-		})
-
-		It("app ls includes new app", func() {
-			apps, err := cli.AppList()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(apps).To(ContainSubstring(appName))
-		})
-
-		It("app show includes app name", func() {
-			appShowOutput, err := cli.AppShow(appName)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(appShowOutput.Name).To(Equal(appName))
-			Expect(appShowOutput.URI).To(BeEmpty())
-		})
-	})
-
-	Context("when creating a new environment", func() {
-		var (
-			testEnvInitErr error
-		)
-		BeforeAll(func() {
-			_, testEnvInitErr = cli.EnvInit(&client.EnvInitRequest{
-				AppName: appName,
-				EnvName: "test",
-				Profile: "default",
-				Prod:    false,
-			})
-		})
-
-		It("env init should succeed", func() {
-			Expect(testEnvInitErr).NotTo(HaveOccurred())
-		})
-	})
-
-	Context("when adding a svc", func() {
-		var (
-			svcInitErr error
-		)
-
-		BeforeAll(func() {
-
-			// Copy the manifest into the copilot/docker-mft folder.
-			err := os.MkdirAll(fmt.Sprintf("./copilot/%s", svcName), 0777)
-			Expect(err).NotTo(HaveOccurred(), fmt.Sprintf("create copilot and %s dirs", svcName))
-
-			destFile, err := os.Create(fmt.Sprintf("./copilot/%s/manifest.yml", svcName))
-			Expect(err).NotTo(HaveOccurred(), "create destination manifest")
-			defer destFile.Close()
-
-			srcFile, err := os.Open(fmt.Sprintf("./%s/manifest.yml", svcName))
-			Expect(err).NotTo(HaveOccurred(), "open source file")
-			defer srcFile.Close()
-
-			_, err = io.Copy(destFile, srcFile)
-			Expect(err).NotTo(HaveOccurred(), "copy manifest file")
-
-			_, svcInitErr = cli.SvcInit(&client.SvcInitRequest{
-				Name:       svcName,
-				SvcType:    "Load Balanced Web Service",
-				Dockerfile: fmt.Sprintf("./%s/Dockerfile", svcName),
-				SvcPort:    "80"})
-
-			It("svc init should succeed", func() {
-				Expect(svcInitErr).NotTo(HaveOccurred())
-			})
-
-			It("manifest should exist", func() {
-				Expect(fmt.Sprintf("./copilot/%s/manifest.yml", svcName)).Should(BeAnExistingFile())
-			})
-		})
-
-		It("svc ls should list the service", func() {
-			svcList, svcListError := cli.SvcList(appName)
-			Expect(svcListError).NotTo(HaveOccurred())
-			Expect(len(svcList.Services)).To(Equal(1))
-
-			svcsByName := map[string]client.SvcDescription{}
-			for _, svc := range svcList.Services {
-				svcsByName[svc.Name] = svc
-			}
-
-			for _, svc := range []string{svcName} {
-				Expect(svcsByName[svc].AppName).To(Equal(appName))
-				Expect(svcsByName[svc].Name).To(Equal(svc))
-			}
-		})
-	})
-
-	Context("when deploying svc", func() {
-		var (
-			appDeployErr error
-			svcInitErr   error
-			initErr      error
-		)
-		BeforeAll(func() {
-			_, appDeployErr = cli.SvcDeploy(&client.SvcDeployInput{
-				Name:     svcName,
-				EnvName:  "test",
-				ImageTag: "princeofthieves",
-			})
-		})
-
-		It("svc deploy should succeed", func() {
-			Expect(appDeployErr).NotTo(HaveOccurred())
-		})
-
-		It("should get the overwritten env var", func() {
-			svc, svcShowErr := cli.SvcShow(&client.SvcShowRequest{
-				AppName: appName,
-				Name:    svcName,
-			})
-			Expect(svcShowErr).NotTo(HaveOccurred())
-			Expect(len(svc.Routes)).To(Equal(1))
-
-			// Make a GET request to the API to retrieve the list of user names from DynamoDB.
-			route := svc.Routes[0]
-			Expect(route.Environment).To(Equal("test"))
-			var resp *http.Response
-			var fetchErr error
-			Eventually(func() (int, error) {
-				resp, fetchErr = http.Get(fmt.Sprintf("%s/magicwords", route.URL))
-				return resp.StatusCode, fetchErr
-			}, "10s", "1s").Should(Equal(200))
-
-			bodyBytes, err := ioutil.ReadAll(resp.Body)
-			Expect(err).NotTo(HaveOccurred())
-
-			type Result struct {
-				Names []string
-			}
-			result := Result{}
-			err = json.Unmarshal(bodyBytes, &result)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(bodyBytes).To(Equal("open sesame"))
-		})
-
-		It("svc logs should display logs", func() {
-			var svcLogs []client.SvcLogsOutput
-			var svcLogsErr error
-			Eventually(func() ([]client.SvcLogsOutput, error) {
-				svcLogs, svcLogsErr = cli.SvcLogs(&client.SvcLogsRequest{
-					AppName: appName,
-					Name:    svcName,
-					EnvName: "test",
-					Since:   "1h",
-				})
-				return svcLogs, svcLogsErr
-			}, "60s", "10s").ShouldNot(BeEmpty())
-
-			for _, logLine := range svcLogs {
-				Expect(logLine.Message).NotTo(Equal(""))
-				Expect(logLine.LogStreamName).NotTo(Equal(""))
-				Expect(logLine.Timestamp).NotTo(Equal(0))
-				Expect(logLine.IngestionTime).NotTo(Equal(0))
-			}
-		})
-			Expect(initErr).NotTo(HaveOccurred())
-		})
-	})
-})
